+++ apps/web/src/app/(with-global-layout)/auctions/write/page.tsx
@@ -1,4 +1,4 @@
-import AuctionFormPage from 'src/features/auction/AuctionFormPage';
+import AuctionFormPage from 'src/features/auction/form/page/AuctionFormPage';
 import type { AuctionMutationPageProps } from 'src/entities/auction/types';
 
 const AuctionMutationPage = async ({ searchParams }: AuctionMutationPageProps) => {
