+++ apps/web/src/features/auction/form/components/AuctionForm.tsx
@@ -17,11 +17,11 @@ import { getExtension } from 'src/entities/auction/utils/extension';
 import { getFormDefaultValues } from 'src/entities/auction/utils/formDefaultValues';
 import { uploadImagesToDB } from 'src/entities/auction/utils/uploadImages';
 import { validateDate } from 'src/entities/auction/utils/validateDate';
-import FormEndDay from 'src/features/auction/FormEndDay';
-import FormEndTime from 'src/features/auction/FormEndTime';
-import FormMaxPoint from 'src/features/auction/FormMaxPoint';
-import FormStartingPoint from 'src/features/auction/FormStartingPoint';
-import ImageUploader from 'src/features/auction/ImageUploader';
+import FormEndDay from 'src/features/auction/form/components/FormEndDay';
+import FormEndTime from 'src/features/auction/form/components/FormEndTime';
+import FormMaxPoint from 'src/features/auction/form/components/FormMaxPoint';
+import FormStartingPoint from 'src/features/auction/form/components/FormStartingPoint';
+import ImageUploader from 'src/features/auction/form/components/ImageUploader';
 import FormDescription from 'src/shared/ui/FormDescription';
 import FormTitle from 'src/shared/ui/FormTitle';
 import PageContainer from 'src/shared/ui/PageContainer';
