+++ apps/web/src/features/auction/form/components/FormStartingPoint.tsx
