+++ apps/web/src/features/auction/list/page/AuctionListPage.tsx
@@ -1,11 +1,11 @@
 import { dehydrate, HydrationBoundary, QueryClient } from '@tanstack/react-query';
 import { prefetchedAuctionList } from 'src/entities/auction/queries/auction';
-import AuctionList from 'src/features/auction/AuctionList';
-import SelectOrder from 'src/features/auction/SelectOrder';
+import AuctionList from 'src/features/auction/list/components/AuctionList';
+import SelectOrder from 'src/features/auction/list/components/SelectOrder';
+import WriteAuctionButton from 'src/features/auction/WriteAuctionButton';
 import PageContainer from 'src/shared/ui/PageContainer';
 import PageTitle from 'src/shared/ui/PageTitle';
 import GoTopButton from 'src/shared/utils/goTopButton';
-import WriteAuctionButton from './WriteAuctionButton';
 import type { AuctionListPageProps } from 'src/entities/auction/types';
 
 const AuctionListPage = async ({ order, keyword }: AuctionListPageProps) => {
